+++ R/modelDiagnostics.R
@@ -14,7 +14,7 @@
 #' Each `RasterBrick` should have the same number of layers,
 #' corresponding to vertical depth slices.
 #'
-#' @details `MESS3D` is a wrapper around `mess` from the `dismo`
+#' @details `MESS3D` is a wrapper around `MESS` from the `modEvA`
 #' package. It calculates MESS for each depth layer. Negative values
 #' indicate areas of extrapolation which should be interpreted with
 #' caution (see Elith *et al*, 2010 in *MEE*).
@@ -27,9 +27,9 @@
 #' The art of modelling range-shifting species.
 #' *Methods in Ecology and Evolution*, 1, 330-342.
 #'
-#' @return A `RasterBrick` of mess scores in each voxel;
-#' layer names correspond to layer names of first
-#' `RasterBrick` in `projection` `list`.
+#' @return A `SpatRaster` vector with MESS scores for each
+#' voxel; layer names correspond to layer names of first
+#' `SpatRaster` vector in `projection` `list`.
 #'
 #' @examples
 #' library(raster)
@@ -74,7 +74,7 @@
 #' messStack <- MESS3D(calibration = calibration, projection = rastList)
 #' plot(messStack)
 #'
-#' @importFrom dismo mess
+#' @importFrom modEvA MESS
 #' @importFrom methods is
 #'
 #' @seealso \code{\link[dismo]{mess}}
@@ -87,8 +87,8 @@ MESS3D <- function (calibration, projection) {
     warning(paste0("'calibration' must be an object of class 'data.frame'.\n"))
     return(NULL)
   }
-  if (!is.list(projection) || !all(unlist(lapply(projection, FUN = function(x) is(x, "RasterBrick"))))) {
-    warning(paste0("'projection' must be a list of 'RasterBrick' objects.\n"))
+  if (!is.list(projection) || !all(unlist(lapply(projection, FUN = function(x) is(x, "SpatRaster"))))) {
+    warning(paste0("'projection' must be a list of 'SpatRaster' objects.\n"))
     return(NULL)
   }
   if(!all(names(projection) %in% colnames(calibration))){
@@ -98,12 +98,16 @@ MESS3D <- function (calibration, projection) {
 
   # Go through the environmental variables and calculate MESS for each layer
   cal <- as.data.frame(calibration[,names(projection)])
-  messStack <- NULL
-  for (i in 1:nlayers(projection[[1]])){
-    proj <- stack(lapply(projection, FUN = function(x) x[[i]]))
-    messLayer <- mess(x = proj, v = cal, full = F)
-    messStack <- stack(c(messStack, messLayer))
+  messStack <- projection
+  for (i in 1:nlyr(projection[[1]])){
+    proj <- c(lapply(projection, FUN = function(x) x[[i]]))
+    projVals <- as.data.frame(proj)
+    projCoords <- terra::crds(proj[[1]])
+    messScores <- MESS(P = projVals, V = cal, verbosity = 0)
+    messLayer <- terra::rasterize(projCoords, proj[[1]], values = messScores["TOTAL"])
+    messStack[[i]] <- messLayer
   }
   names(messStack) <- names(projection[[1]])
+  messStack <- rast(messStack)
   return(messStack)
 }