+++ R/xyzSample.R
@@ -1,21 +1,21 @@
-#' @title Sampling from rasterBrick using 3D coordinates
+#' @title Sampling from a `SpatRaster` vector using 3D coordinates
 #'
 #' @description Gets values at x,y,z occurrences from a
 #' given 3D environmental variable brick
 #'
-#' @param occs A dataframe with at least three columns
+#' @param occs A `data.frame` with at least three columns
 #' named "longitude", "latitude", and "depth", or that
 #' can be coerced into this format.
 #'
-#' @param envBrick A `rasterBrick` object with
+#' @param envBrick A `SpatRaster` vector object with
 #' one environmental variable. Each layer represents
 #' a depth slice. See Details for more information.
 #'
-#' @param verbose `logical`. Switching to `FALSE` mutes message describing
-#' which columns in `occs1` and `occs2` are interpreted as x, y, and z
-#' coordinates.
+#' @param verbose `logical`. Switching to `FALSE` mutes message
+#' describing which columns in `occs1` and `occs2` are interpreted
+#' as x, y, and z coordinates.
 #'
-#' @details The `envBrick` `rasterBrick` object should
+#' @details The `SpatRaster` vector object should
 #' have numeric names that correspond with the beginning
 #' depth of a particular depth slice. For example, one
 #' might have three layers, one from 0 to 10m, one from
@@ -32,24 +32,24 @@
 #' to number of rows of input `occs` `data.frame`.
 #'
 #' @examples
-#' library(raster)
+#' library(terra)
 #'
 #' # Create test raster
-#' r1 <- raster(ncol=10, nrow=10)
+#' r1 <- rast(ncol=10, nrow=10)
 #' values(r1) <- 1:100
-#' r2 <- raster(ncol=10, nrow=10)
+#' r2 <- rast(ncol=10, nrow=10)
 #' values(r2) <- c(rep(20, times = 50), rep(60, times = 50))
-#' r3 <- raster(ncol=10, nrow=10)
+#' r3 <- rast(ncol=10, nrow=10)
 #' values(r3) <- 8
-#' envBrick <- brick(r1, r2, r3)
+#' envBrick <- c(r1, r2, r3)
 #' names(envBrick) <- c(0, 10, 30)
 #'
 #' # Create test occurrences
 #' set.seed(0)
-#' longitude <- sample(extent(envBrick)[1]:extent(envBrick)[2],
+#' longitude <- sample(ext(envBrick)[1]:ext(envBrick)[2],
 #'                     size = 10, replace = FALSE)
 #' set.seed(0)
-#' latitude <- sample(extent(envBrick)[3]:extent(envBrick)[4],
+#' latitude <- sample(ext(envBrick)[3]:ext(envBrick)[4],
 #'                    size = 10, replace = FALSE)
 #' set.seed(0)
 #' depth <- sample(0:35, size = 10, replace = TRUE)
@@ -61,7 +61,6 @@
 #' # How to use
 #' occurrences$envtValue <- occSample3d
 #'
-#' @import raster
 #' @importFrom terra rast extract
 #'
 #' @keywords dataPrep
@@ -79,12 +78,10 @@ xyzSample <- function(occs, envBrick, verbose = TRUE){
     return(NULL)
   }
 
-  if(!inherits(envBrick, what = "RasterBrick")){
-    warning(paste0("'envBrick' must be of class 'RasterBrick'.\n"))
+  if(!inherits(envBrick, what = "SpatRaster")){
+    warning(paste0("'envBrick' must be of class 'SpatRaster'.\n"))
     return(NULL)
   }
-
-  envBrick <- rast(envBrick)
 
   if (!is.logical(verbose)) {
     warning(message("Argument 'verbose' is not of type 'logical'.\n"))