+++ tests/testthat/test-inputProcessingUtilties.R
@@ -1,21 +1,21 @@
-library(raster)
+library(terra)
 library(sp)
 
 # voluModel:::columnParse() tests ----
 # Create sample raster
-r <- raster(ncol=10, nrow=10)
+r <- rast(ncol=10, nrow=10)
 values(r) <- 1:100
 
 # Create sample raster brick
-rBrick <- brick(r, r*10, r*100)
+rBrick <- c(r, r*10, r*100)
 names(rBrick) <- c(0, 10, 100)
 
 # Create test occurrences
 set.seed(0)
-longitude <- sample(extent(rBrick)[1]:extent(rBrick)[2],
+longitude <- sample(ext(rBrick)[1]:ext(rBrick)[2],
                     size = 10, replace = FALSE)
 set.seed(0)
-latitude <- sample(extent(rBrick)[3]:extent(rBrick)[4],
+latitude <- sample(ext(rBrick)[3]:ext(rBrick)[4],
                    size = 10, replace = FALSE)
 set.seed(0)
 depth <- sample(0:98, size = 10, replace = TRUE)
@@ -89,15 +89,15 @@ test_that("columnParse outputs as expected", {
 
 # downsample() tests ----
 # Create sample raster
-r <- raster(ncol=5, nrow=5)
+r <- rast(ncol=5, nrow=5)
 values(r) <- 1:25
 
 # Create test occurrences
 set.seed(0)
-longitude <- sample(extent(r)[1]:extent(r)[2],
+longitude <- sample(ext(r)[1]:ext(r)[2],
                     size = 5, replace = TRUE)
 set.seed(0)
-latitude <- sample(extent(r)[3]:extent(r)[4],
+latitude <- sample(ext(r)[3]:ext(r)[4],
                    size = 5, replace = TRUE)
 occurrences <- as.data.frame(cbind(longitude,latitude))