+++ tests/testthat/test-xyzSample.R
@@ -1,19 +1,19 @@
-library(raster)
+library(terra)
 
 # Create sample raster
-r <- raster(ncol=10, nrow=10)
+r <- rast(ncol=10, nrow=10)
 values(r) <- 1:100
 
 # Create sample raster brick
-rBrick <- brick(r, r*10, r*100)
+rBrick <- c(r, r*10, r*100)
 names(rBrick) <- c(0, 10, 100)
 
 # Create test occurrences
 set.seed(0)
-longitude <- sample(extent(rBrick)[1]:extent(rBrick)[2],
+longitude <- sample(ext(rBrick)[1]:ext(rBrick)[2],
                     size = 10, replace = FALSE)
 set.seed(0)
-latitude <- sample(extent(rBrick)[3]:extent(rBrick)[4],
+latitude <- sample(ext(rBrick)[3]:ext(rBrick)[4],
                    size = 10, replace = FALSE)
 set.seed(0)
 depth <- sample(0:98, size = 10, replace = TRUE)
@@ -42,7 +42,7 @@ test_that("xyzSample returns appropriate information", {
   # Test function
   occSample3d <- xyzSample(occurrences, rBrick)
 
-  expect_type(occSample3d, "double")
+  expect_type(occSample3d, "integer")
   expect_length(occSample3d, 10)
   expect_equal(0, sum(is.na(occSample3d)))
 })
@@ -52,7 +52,7 @@ test_that("xyzSample column parsing works", {
   occurrences$"z" <- occurrences$depth
   occurrences$"zebra" <- occurrences$depth
   occSample3d <- xyzSample(occurrences, rBrick)
-  expect_type(occSample3d, "double")
+  expect_type(occSample3d, "integer")
   expect_length(occSample3d, 10)
   expect_equal(0, sum(is.na(occSample3d)))
 })